fix: replace `long` with `long long` for compatibility on windows

+++ inc/cmd/list.h
@@ -7,6 +7,6 @@
 int list_cmd(int argc, char *argv[]);
 void print_headers(PadStudent *pad);
 Vec *get_students_by_property(Vec* sv, const char *property, const char *value);
-Vec *get_students_in_range(Vec* sv, const char *property, long start, long end);
+Vec *get_students_in_range(Vec* sv, const char *property, long long start, long long end);
 
 #endif
